# swagger/SocialnetApiDocs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Socialnet API",
    "description": "Code Lounge",
    "version": "1.6"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/post/{postId}/comment": {
      "get": {
        "tags": [
          "comment-controller"
        ],
        "summary": "Получение комментариев по посту",
        "operationId": "getComments",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "ID поста",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Комментарии успешно получены",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDtoCommentDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": [
          "comment-controller"
        ],
        "summary": "Обновление комментария",
        "operationId": "updateComment",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "ID поста",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Комментарий успешно обновлен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "comment-controller"
        ],
        "summary": "Создание комментария",
        "operationId": "createComment",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "ID поста",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "description": "Текст комментария"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Комментарий успешно создан",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/post/{postId}/comment/{commentId}/subcomment": {
      "get": {
        "tags": [
          "comment-controller"
        ],
        "summary": "Получение субкомментариев по комментарию",
        "operationId": "getSubComments",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "ID поста",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "commentId",
            "in": "path",
            "description": "ID комментария",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Субкомментарии успешно получены",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDtoCommentDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/post/{postId}/comment/{commentId}": {
      "delete": {
        "tags": [
          "comment-controller"
        ],
        "summary": "Удаление комментария",
        "operationId": "deleteComment",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "ID поста",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "commentId",
            "in": "path",
            "description": "ID комментария",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Комментарий успешно удален",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommentUpdateDto": {
        "type": "object",
        "properties": {
          "commentText": {
            "type": "string",
            "description": "Текст комментария"
          },
          "id": {
            "type": "string",
            "description": "ID комментария",
            "format": "uuid"
          }
        },
        "description": "Данные для обновления комментария"
      },
      "CommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID комментария",
            "format": "uuid"
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Удален ли комментарий"
          },
          "commentType": {
            "type": "string",
            "description": "Тип публикации",
            "example": "POST",
            "enum": [
              "COMMENT",
              "POST"
            ]
          },
          "time": {
            "type": "string",
            "description": "Время комментария",
            "format": "date-time"
          },
          "timeChanged": {
            "type": "string",
            "description": "Время изменения комментария",
            "format": "date-time"
          },
          "authorId": {
            "type": "string",
            "description": "ID автора",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "description": "ID родительского комментария",
            "format": "uuid"
          },
          "commentText": {
            "type": "string",
            "description": "Текст комментария"
          },
          "postId": {
            "type": "string",
            "description": "ID поста",
            "format": "uuid"
          },
          "isBlocked": {
            "type": "boolean",
            "description": "Заблокирован ли комментарий"
          },
          "likeAmount": {
            "type": "integer",
            "description": "Количество лайков",
            "format": "int32"
          },
          "myLike": {
            "type": "boolean",
            "description": "Флаг авторства лайка"
          },
          "commentsCount": {
            "type": "integer",
            "description": "Количество комментариев",
            "format": "int32"
          },
          "imagePath": {
            "type": "string",
            "description": "Путь к изображению"
          }
        },
        "description": "Данные комментария"
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PageableDto": {
        "type": "object",
        "properties": {
          "sort": {
            "$ref": "#/components/schemas/SortDto"
          },
          "pageNumber": {
            "type": "integer",
            "description": "Номер страницы",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "Размер страницы",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "description": "Смещение",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean",
            "description": "Без разбивки на страницы"
          },
          "paged": {
            "type": "boolean",
            "description": "Разбито на страницы"
          }
        },
        "description": "Данные для постраничного вывода"
      },
      "SortDto": {
        "type": "object",
        "properties": {
          "unsorted": {
            "type": "boolean",
            "description": "Не отсортировано"
          },
          "sorted": {
            "type": "boolean",
            "description": "Отсортировано"
          },
          "empty": {
            "type": "boolean",
            "description": "Пусто"
          }
        },
        "description": "Данные сортировки"
      },
      "PageResponseDtoCommentDto": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "description": "Общее количество элементов",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "description": "Общее количество страниц",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "description": "Номер страницы",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "description": "Размер страницы",
            "format": "int32"
          },
          "first": {
            "type": "boolean",
            "description": "Первая страница"
          },
          "last": {
            "type": "boolean",
            "description": "Последняя страница"
          },
          "numberOfElements": {
            "type": "integer",
            "description": "Количество элементов на странице",
            "format": "int32"
          },
          "empty": {
            "type": "boolean",
            "description": "Пустая страница"
          },
          "sort": {
            "$ref": "#/components/schemas/SortDto"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableDto"
          },
          "content": {
            "minItems": 0,
            "uniqueItems": true,
            "type": "array",
            "description": "Содержимое страницы",
            "items": {
              "$ref": "#/components/schemas/CommentDto"
            }
          }
        },
        "description": "Ответ с данными страницы"
      },
...
    }
  }
}